+++ ReCaptcha/Model/ConfigInterface.php
@@ -16,26 +16,4 @@ use Magento\Framework\Phrase;
  */
 interface ConfigInterface
 {
-    /**
-     * Get google recaptcha public key
-     * @return string
-     */
-    public function getPublicKey(): string;
-
-    /**
-     * Get google recaptcha private key
-     * @return string
-     */
-    public function getPrivateKey(): string;
-
-    /**
-     * @return bool
-     */
-    public function isInvisibleRecaptcha(): bool;
-
-    /**
-     * Get reCaptcha type
-     * @return string
-     */
-    public function getType(): string;
 }
