--- ReCaptcha/etc/config.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
- -->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Store:etc/config.xsd">
-    <default>
-        <recaptcha>
-            <general>
-                <type>recaptcha_v3</type>
-            </general>
-        </recaptcha>
-    </default>
-</config>
