+++ ReCaptcha/etc/di.xml
@@ -7,20 +7,5 @@
  -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <preference for="Magento\ReCaptcha\Model\ConfigInterface" type="Magento\ReCaptcha\Model\Config"/>
     <preference for="Magento\ReCaptcha\Model\ValidateInterface" type="Magento\ReCaptcha\Model\Validate"/>
-
-    <!-- Set sensitive information -->
-    <type name="Magento\Config\Model\Config\TypePool">
-        <arguments>
-            <argument name="sensitive" xsi:type="array">
-                <item name="recaptcha/general/public_key" xsi:type="string">1</item>
-                <item name="recaptcha/general/private_key" xsi:type="string">1</item>
-            </argument>
-            <argument name="environment" xsi:type="array">
-                <item name="recaptcha/general/public_key" xsi:type="string">1</item>
-                <item name="recaptcha/general/private_key" xsi:type="string">1</item>
-            </argument>
-        </arguments>
-    </type>
 </config>
