+++ ReCaptchaAdminUi/etc/adminhtml/system.xml
@@ -70,39 +70,51 @@
             <group id="frontend" translate="label" type="text" sortOrder="30" showInDefault="1" showInWebsite="1"
                    showInStore="1">
                 <label>Frontend</label>
-                <field id="enabled" translate="label" type="select" sortOrder="40" showInDefault="1" showInWebsite="1"
-                       showInStore="0">
-                    <label>Enable</label>
-                    <source_model>Magento\Config\Model\Config\Source\Yesno</source_model>
-                </field>
-                <field id="position" translate="label" type="select" sortOrder="60" showInDefault="1" showInWebsite="1"
+
+                <field id="type" translate="label" type="select" sortOrder="10" showInDefault="1" showInWebsite="1"
                        showInStore="0" canRestore="1">
-                    <label>Invisible badge position</label>
-                    <source_model>Magento\ReCaptchaAdminUi\Model\Config\Source\Position</source_model>
-                    <depends>
-                        <field id="enabled">1</field>
-                    </depends>
-                    <comment>Only for Invisible ReCaptcha v2 or ReCaptcha v3.</comment>
+                    <label>reCaptcha type</label>
+                    <source_model>Magento\ReCaptchaAdminUi\Model\Config\Source\Type</source_model>
                 </field>
-                <field id="min_score" translate="label" type="text" sortOrder="60" showInDefault="1" showInWebsite="1"
-                       showInStore="0" canRestore="1">
-                    <label>Minimum score</label>
-                    <depends>
-                        <field id="enabled">1</field>
-                        <field id="recaptcha/general/type">recaptcha_v3</field>
-                    </depends>
-                    <comment>From 0.0 to 1.0, where 0.0 is absolutely a robot and 1.0 is a human.</comment>
+
+                <field id="public_key" translate="label comment" type="text" sortOrder="20" showInDefault="1" showInWebsite="1"
+                       showInStore="0" canRestore="0">
+                    <label>Google API website key</label>
                 </field>
-                <field id="theme" translate="label" type="select" sortOrder="60" showInDefault="1" showInWebsite="1"
+
+                <field id="private_key" translate="label comment" type="password" sortOrder="30" showInDefault="1" showInWebsite="1"
+                       showInStore="0" canRestore="0">
+                    <label>Google API secret key</label>
+                </field>
+
+                <field id="theme" translate="label" type="select" sortOrder="40" showInDefault="1" showInWebsite="1"
                        showInStore="0" canRestore="1">
                     <label>Theme</label>
                     <source_model>Magento\ReCaptchaAdminUi\Model\Config\Source\Theme</source_model>
                     <depends>
-                        <field id="enabled">1</field>
-                        <field id="recaptcha/general/type">recaptcha</field>
+                        <field id="recaptcha/frontend/type">recaptcha</field>
                     </depends>
                 </field>
-                <field id="lang" translate="label" type="text" sortOrder="65" showInDefault="1" showInWebsite="1"
+
+                <field id="size" translate="label" type="select" sortOrder="50" showInDefault="1" showInWebsite="1"
+                       showInStore="0" canRestore="1">
+                    <label>Size</label>
+                    <source_model>Magento\ReCaptchaAdminUi\Model\Config\Source\Size</source_model>
+                    <depends>
+                        <field id="recaptcha/frontend/type">recaptcha</field>
+                    </depends>
+                </field>
+
+                <field id="min_score" translate="label" type="text" sortOrder="60" showInDefault="1" showInWebsite="1"
+                       showInStore="0" canRestore="1">
+                    <label>Minimum score</label>
+                    <depends>
+                        <field id="recaptcha/general/type">recaptcha_v3</field>
+                    </depends>
+                    <comment>From 0.0 to 1.0, where 0.0 is absolutely a robot and 1.0 is a human.</comment>
+                </field>
+
+                <field id="lang" translate="label" type="text" sortOrder="70" showInDefault="1" showInWebsite="1"
                        showInStore="1" canRestore="1">
                     <label>Language Code</label>
                     <comment><![CDATA[
@@ -111,18 +123,13 @@
                         See <strong><a target="_blank" href="https://developers.google.com/recaptcha/docs/language">supported Language codes</a></strong>.
                     </div>
                 ]]></comment>
-                    <depends>
-                        <field id="enabled">1</field>
-                    </depends>
                 </field>
-                <field id="size" translate="label" type="select" sortOrder="70" showInDefault="1" showInWebsite="1"
+
+                <field id="position" translate="label" type="select" sortOrder="80" showInDefault="1" showInWebsite="1"
                        showInStore="0" canRestore="1">
-                    <label>Size</label>
-                    <source_model>Magento\ReCaptchaAdminUi\Model\Config\Source\Size</source_model>
-                    <depends>
-                        <field id="enabled">1</field>
-                        <field id="recaptcha/general/type">recaptcha</field>
-                    </depends>
+                    <label>Invisible badge position</label>
+                    <source_model>Magento\ReCaptchaAdminUi\Model\Config\Source\Position</source_model>
+                    <comment>Only for Invisible ReCaptcha v2 or ReCaptcha v3.</comment>
                 </field>
             </group>
         </section>
