+++ ReCaptchaFrontendUi/etc/di.xml
@@ -9,4 +9,17 @@
         xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <preference for="Magento\ReCaptchaFrontendUi\Model\FrontendConfigInterface"
                 type="Magento\ReCaptchaFrontendUi\Model\FrontendConfig"/>
+
+    <type name="Magento\Config\Model\Config\TypePool">
+        <arguments>
+            <argument name="sensitive" xsi:type="array">
+                <item name="recaptcha/frontend/public_key" xsi:type="string">1</item>
+                <item name="recaptcha/frontend/private_key" xsi:type="string">1</item>
+            </argument>
+            <argument name="environment" xsi:type="array">
+                <item name="recaptcha/frontend/public_key" xsi:type="string">1</item>
+                <item name="recaptcha/frontend/private_key" xsi:type="string">1</item>
+            </argument>
+        </arguments>
+    </type>
 </config>
