+++ ReCaptchaNewsletter/Model/IsEnabledForNewsletter.php
@@ -8,7 +8,6 @@ declare(strict_types=1);
 namespace Magento\ReCaptchaNewsletter\Model;
 
 use Magento\Framework\App\Config\ScopeConfigInterface;
-use Magento\ReCaptcha\Model\{ConfigInterface as ReCaptchaConfig};
 use Magento\ReCaptchaFrontendUi\Model\ConfigEnabledInterface;
 use Magento\ReCaptchaFrontendUi\Model\FrontendConfigInterface as ReCaptchaFrontendUiConfig;
 use Magento\Store\Model\ScopeInterface;
@@ -21,11 +20,6 @@ class IsEnabledForNewsletter implements IsEnabledForNewsletterInterface, ConfigE
     private const XML_PATH_ENABLED_FOR_NEWSLETTER = 'recaptcha/frontend/enabled_for_newsletter';
 
     /**
-     * @var ReCaptchaConfig
-     */
-    private $reCaptchaConfig;
-
-    /**
      * @var ReCaptchaFrontendUiConfig
      */
     private $reCaptchaFrontendConfig;
@@ -36,16 +30,13 @@ class IsEnabledForNewsletter implements IsEnabledForNewsletterInterface, ConfigE
     private $scopeConfig;
 
     /**
-     * @param ReCaptchaConfig $reCaptchaConfig
      * @param ReCaptchaFrontendUiConfig $reCaptchaFrontendConfig
      * @param ScopeConfigInterface $scopeConfig
      */
     public function __construct(
-        ReCaptchaConfig $reCaptchaConfig,
         ReCaptchaFrontendUiConfig $reCaptchaFrontendConfig,
         ScopeConfigInterface $scopeConfig
     ) {
-        $this->reCaptchaConfig = $reCaptchaConfig;
         $this->reCaptchaFrontendConfig = $reCaptchaFrontendConfig;
         $this->scopeConfig = $scopeConfig;
     }
@@ -55,7 +46,9 @@ class IsEnabledForNewsletter implements IsEnabledForNewsletterInterface, ConfigE
      */
     public function isEnabled(): bool
     {
-        if (!$this->reCaptchaFrontendConfig->isFrontendEnabled() || !$this->reCaptchaConfig->isInvisibleRecaptcha()) {
+        if (!$this->reCaptchaFrontendConfig->isFrontendEnabled()
+            || !$this->reCaptchaFrontendConfig->isInvisibleRecaptcha()
+        ) {
             return false;
         }
